Add files via upload
In this updated version of Hist-unmix we modified the code to deal with possible duplicate values that could arise from cubic interpolations. An option of choosing a different value for the Monte Carlo error propagation (other than a reduced chi-squared) was also included.  To avoid incompatibilities and to guarantee Hist-unmix's functionality, we recommend the creation of an environment (as described in the README file).